Move `Guard` and associated bits out to a separate file.
`compile/mod.rs` is mostly -- and, in my opinion, sensibly -- trait
definitions and the like, but `Guard` has quite a bit of behaviour to it
that feels a bit out of place. By putting `GuardIdx` in there we also
make it easy for people to access its `.0` field.

This commit thus moves that code out into a separate file. This was
going to be part of a larger PR to make `Guard`s more like `Location`s,
using atomics etc. but having done a bit of that work I realise it's
almost certainly premature optimisation, and our effort (and complexity
budget) is better spent elsewhere. This part still seems useful though.
